Fixed design do form de gestão de restaurantes. Fixed bug onde os dados não eram atualizados na datagridview. Alterado o tipo de dados de Pessoa_Trabalhador(salario) int32 -> decimal

+++ RestGest/Trabalhador.cs
@@ -14,7 +14,7 @@ namespace RestGest
     
     public partial class Trabalhador : Pessoa
     {
-        public int Salario { get; set; }
+        public decimal Salario { get; set; }
         public string Posicao { get; set; }
         public int RestauranteId { get; set; }
         public bool IsDeleted { get; set; }
